+++ CHANGES.md
@@ -6,6 +6,7 @@
    * Added `globalStyles.whiteSpaceStyle` setting to color schemes.
 
 ### Improvements
+ * improved layout of the Setup Dialog shown when the application is started for the first time.
 
 ### Fixes
  * Save As now selects correct file extension by default.